move redundant _close, _equals to one file
